services/auth-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
psycopg[binary]>=3.2.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.0.0
slowapi==0.1.9
python-json-logger==2.0.7

